implemented user dao

+++ src/main/java/api/UserDao.java
@@ -13,6 +13,6 @@ public interface UserDao {
     User getUserByLogin(String login) throws IOException;
     User getUserById(Integer id) throws IOException;
     void removeUserByLogin(String login) throws IOException;
-    void removeUserById(Long id) throws IOException;
+    void removeUserById(Integer id) throws IOException;
 
 }
